+++ test-java-lib/src/main/java/com/uber/lib/unannotated/LambdaBox.java
@@ -1,0 +1,4 @@
+package com.uber.lib.unannotated;
+
+/* @NullMarked */
+public class LambdaBox<T> {}
