+++ test-java-lib/src/main/java/com/uber/lib/unannotated/LambdaConsumer.java
@@ -1,0 +1,6 @@
+package com.uber.lib.unannotated;
+
+/* @NullMarked */
+public interface LambdaConsumer<T> {
+  void accept(T value);
+}
